:recycle: Separate classes by files
