+++ 06.bowling_object/frame.rb
@@ -1,0 +1,15 @@
+class Frame
+  attr_reader :shots
+
+  def initialize(shots)
+    @shots = shots
+  end
+
+  def sum
+    @shots.sum
+  end
+
+  def strike?
+    @shots[0] == 10
+  end
+end
